+++ scripts/release.mjs
@@ -1,8 +1,12 @@
-const fs = require("fs");
-const process = require("process");
-const archiver = require("archiver");
-const path = require("path");
-const { releaseDirectories, releaseFiles, updateFile } = require("./commons");
+import fs from "fs";
+import process from "process";
+import archiver from "archiver";
+import { fileURLToPath } from "url";
+import path, { dirname } from "path";
+import { releaseDirectories, releaseFiles, updateFile } from "./commons.mjs";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 const tarballReleaseFiles = () =>
   new Promise((resolve, reject) => {